Adding extremely rudimentary "kind of" support
At some point the plan is to add full on matpath.  Sadly I don't see a
way for matpath to be effectively used for relationships.  It's great if
you want to get all of an objects children or parents, but if you want
to say, join on an objects parents (my use case here) you need to do
string manipulation in the database and that is not a simple task...